+++ src/engine.ts
@@ -1,49 +1,97 @@
+import fs from 'fs'
 import path from 'path'
+import url from 'url'
+import type { Marp } from '@marp-team/marp-core'
 import { Marpit } from '@marp-team/marpit'
-import importFrom from 'import-from'
+import { resolve as importMetaResolve } from 'import-meta-resolve'
 import { pkgUp } from 'pkg-up'
 import { error } from './error'
 
-export type Engine = typeof Marpit
-export type ResolvableEngine = Engine | DelayedEngineResolver | string
+/** @internal */
+export const _silentImport = async <T = any>(
+  moduleId: string,
+  from?: string
+): Promise<T | null> => {
+  const basePath = path.join(from || process.cwd(), '_.js')
+  const dirPath = path.dirname(basePath)
+  const moduleFilePath = path.resolve(dirPath, moduleId)
 
-const delayedEngineResolverSymbol = Symbol('delayedEngineResolver')
+  try {
+    const stat = await fs.promises.stat(moduleFilePath)
 
-type DelayedEngineResolver = {
-  [delayedEngineResolverSymbol]: () => Promise<Engine>
+    if (stat.isFile()) moduleId = url.pathToFileURL(moduleFilePath).toString()
+  } catch (e: unknown) {
+    // No ops
+  }
+
+  try {
+    const resolved = await importMetaResolve(
+      moduleId,
+      url.pathToFileURL(basePath).toString()
+    )
+
+    // Try to import without `file:` protocol first
+    if (resolved.startsWith('file:')) {
+      try {
+        return await import(url.fileURLToPath(resolved))
+
+        // NOTE: Fallback cannot test because of overriding `import` in Jest context.
+        /* c8 ignore start */
+      } catch (e) {
+        /* fallback */
+      }
+    }
+
+    return await import(resolved)
+    /* c8 ignore stop */
+  } catch (e) {
+    return null
+  }
 }
 
-const delayedEngineResolver = (
-  fn: () => Promise<Engine>
-): DelayedEngineResolver => ({ [delayedEngineResolverSymbol]: fn })
+type MarpitInstanceOrClass = Marpit | typeof Marpit
 
-export class ResolvedEngine {
-  klass: Engine
+export type Engine = MarpitInstanceOrClass | FunctionEngine
+
+type FunctionEngine = (
+  opts?: Marpit.Options
+) => MarpitInstanceOrClass | Promise<MarpitInstanceOrClass>
+
+export type ResolvableEngine = Engine | string
+
+const preResolveAsyncSymbol = Symbol('preResolveAsync')
+
+export class ResolvedEngine<T extends Engine = Engine> {
+  klass: T
 
   private _cachedPackage?: Record<string, any> | null
 
-  private static _defaultEngine: ResolvedEngine | undefined
+  private static _defaultEngine: ResolvedEngine<typeof Marp> | undefined
 
-  static async resolve(
+  static async resolve<T extends Engine = Engine>(
     engine: ResolvableEngine | ResolvableEngine[],
     from?: string
-  ): Promise<ResolvedEngine> {
+  ): Promise<ResolvedEngine<T>> {
     return new ResolvedEngine(await ResolvedEngine.resolveModule(engine, from))
   }
 
-  static async resolveDefaultEngine(): Promise<ResolvedEngine> {
+  static async resolveDefaultEngine(): Promise<ResolvedEngine<typeof Marp>> {
     if (
       ResolvedEngine._defaultEngine === undefined ||
       process.env.NODE_ENV === 'test'
     ) {
       ResolvedEngine._defaultEngine = await ResolvedEngine.resolve([
-        '@marp-team/marp-core', // Manually-installed Marp Core
+        // Manually-installed Marp Core
+        '@marp-team/marp-core',
 
         // Bundled Marp Core
-        // eslint-disable-next-line @typescript-eslint/no-var-requires -- import statement brings segmentation fault: https://github.com/marp-team/marp-cli/issues/487
-        delayedEngineResolver(async () => require('@marp-team/marp-core').Marp),
+        Object.assign(
+          () => import('@marp-team/marp-core').then(({ Marp }) => Marp),
+          { [preResolveAsyncSymbol]: true }
+        ),
       ])
     }
+
     return ResolvedEngine._defaultEngine
   }
 
@@ -63,16 +111,23 @@ export class ResolvedEngine {
     for (const eng of ([] as ResolvableEngine[]).concat(engine)) {
       if (typeof eng === 'string') {
         resolved =
-          (from && importFrom.silent(path.dirname(path.resolve(from)), eng)) ||
-          importFrom.silent(process.cwd(), eng)
-
-        if (resolved && typeof resolved === 'object' && 'default' in resolved) {
-          resolved = resolved.default
-        }
-      } else if (typeof eng === 'object' && eng[delayedEngineResolverSymbol]) {
-        resolved = await eng[delayedEngineResolverSymbol]()
+          (from &&
+            (await _silentImport(eng, path.dirname(path.resolve(from))))) ||
+          (await _silentImport(eng))
+      } else if (typeof eng === 'function' && eng[preResolveAsyncSymbol]) {
+        resolved = await (eng as any)()
       } else {
         resolved = eng
+      }
+
+      // Resolve default export
+      while (
+        resolved &&
+        typeof resolved === 'object' &&
+        'default' in resolved &&
+        resolved !== resolved.default
+      ) {
+        resolved = resolved.default
       }
 
       if (resolved) break
@@ -82,7 +137,7 @@ export class ResolvedEngine {
     return resolved
   }
 
-  private constructor(klass: Engine) {
+  private constructor(klass: T) {
     this.klass = klass
   }