+++ test/_configs/custom-engine/import.js
@@ -1,0 +1,1 @@
+module.exports = { engine: import('./custom-engine.js') }